Introduce an option to enable monitoring the parent process
Not all AppHost processes will want this behaviour, so add an option so
that the launcher can specify whether or not to enable the behaviour.

+++ RedGate.AppHost.Client/Program.cs
@@ -24,9 +24,14 @@ namespace RedGate.AppHost.Client
                 options.Debug = true;
 #endif
                 if (options.Debug)
+                {
                     ConsoleNativeMethods.AllocConsole();
+                }
 
-                new ParentProcessMonitor(() => { Process.GetCurrentProcess().Kill(); }).Start();
+                if (options.MonitorParentProcess)
+                {
+                    new ParentProcessMonitor(() => { Process.GetCurrentProcess().Kill(); }).Start();
+                }
 
                 MainInner(options.ChannelId, options.Assembly);
             }
